+++ Rona/Rona_Module.php
@@ -5,7 +5,7 @@
  * @copyright Copyright (c) 2018 Ryan Whitman (https://ryanwhitman.com)
  * @license https://opensource.org/licenses/MIT
  * @link https://github.com/RyanWhitman/ronaphp
- * @version 1.0.0
+ * @version 1.001.0
  */
 
 namespace Rona;
@@ -18,7 +18,7 @@ class Rona_Module extends Module {
 	/**
 	 * @see parent class
 	 */
-	protected function register_param_filter_groups() {
+	public function register_param_filter_groups() {
 		$this->register_param_filter_group('general', '\Rona\Rona_Param_Filters');
 	}
 }