+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### v. 1.001.0
+
+- Previously, module components such as resources and routes were registered in the module's __construct() method. Now, all components except for the module's configuration are registered in a new init() method housed in the Rona class. The init() method is automatically executed in the run() method. This allows all modules to be registered before registering their components. To accommodate this change, several existing methods were changed from protected to public.
+- A new locate_param_filter() method was added to the Rona class. This method finds a param filter by either its name or module/name combo.
+- A new copy() method has been added to the Param_Filter_Group class that allows existing filters to be copied. Copying a filter allows the developer to maintain the existing callback, but modify the default options.
+- Added editor config.
+
 ### v. 1.0.0
 
 - The pre-built Rona param filters were previously stored in a separate repo. That repo. has been deleted and the Rona module is now built into the RonaPHP framework.